[TeamC] lots of UI fixes/edits and updating test
