+++ Team-C/test_wizard/lib/views/view_test_view.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:test_wizard/widgets/tw_app_bar.dart';
-import 'package:test_wizard/widgets/view_test.dart';
+import 'package:test_wizard/widgets/view_assessment_page.dart';
 
 class CreateViewTest extends StatelessWidget {
   const CreateViewTest({super.key});
@@ -11,6 +11,7 @@ class CreateViewTest extends StatelessWidget {
       appBar: TWAppBar(
         context: context,
         screenTitle: 'View Quiz',
+        implyLeading: true,
       ),
       body: const ViewAssessmentPage(),
     );
